+++ package.json
@@ -29,7 +29,8 @@
     "react-leaflet": "^5.0.0",
     "react-router-dom": "^7.8.1",
     "regenerator-runtime": "^0.14.1",
-    "sass": "^1.90.0"
+    "sass": "^1.90.0",
+    "split-type": "^0.3.4"
   },
   "devDependencies": {
     "@eslint/js": "^9.32.0",
